Update gui thong tin

+++ src/helper/commonFunctions.js
@@ -6,20 +6,21 @@ const axios = require('axios');
 const globalConfig = require("../config/global");
 const fs = require('fs')
 var ip = require("ip");
+const osu = require('node-os-utils')
 
 async function exists(path) {
   try {
     return await fs.existsSync(path)
-  } catch(ex) {
+  } catch (ex) {
     console.error(ex);
     return false
   }
 }
 
- function getIp(){
-  try{
-    return  ip.address();
-  }catch(ex){
+function getIp() {
+  try {
+    return ip.address();
+  } catch (ex) {
     return ex;
   }
 }
@@ -27,13 +28,13 @@ async function exists(path) {
 
 module.exports = {
   async getProp(key) {
-    try{
+    try {
       const { stdout, stderr } = await util.promisify(exec)(`getprop ${key}`);
       if (stderr) {
         return stderr.trim();
       }
       return stdout.trim();
-    }catch(ex){
+    } catch (ex) {
       return ex.stderr.trim();
     }
   },
@@ -55,46 +56,72 @@ module.exports = {
         const data = await axios.get(`${globalConfig.apiURL}/user/get/${username}`)
         if (data.data.data) {
           userConfig = data.data.data;
-          //Sinh them thong tin thiet bi
-          userConfig.deviceId = require('shortid').generate();
-          userConfig.deviceName = await this.getUserSSH();
-
           await fs.promises.writeFile(globalConfig.configPath, JSON.stringify(userConfig));
           return userConfig;
         } else {
           console.log("User khong ton tai!");
         }
-
       }
-    }else{
+    } else {
       let rawdata = fs.readFileSync(globalConfig.configPath);
       userConfig = JSON.parse(rawdata);
       return userConfig;
     }
   },
-  async removeUserConfig(){
-    try
-    {
-      if (await exists(globalConfig.configPath)){
-        await fs.unlinkSync(globalConfig.configPath)
+  async getDeviceConfig() {
+    let deviceConfig;
+    if (await exists(globalConfig.deviceConfig)) {
+      let rawdata = fs.readFileSync(globalConfig.deviceConfig);
+      deviceConfig = JSON.parse(rawdata);
+    } else {
+      deviceConfig = {
+        deviceId: require("shortid").generate(),
+        deviceName: await this.getProp("ro.product.model"),
+        sshUser: await this.getUserSSH(),
+        localIp: this.getDeviceIP(),
+        cpuUse: 0,
+        cpuCores: this.getCpuCores(),
+        model: await this.getProp("ro.product.model"),
       }
-      console.log("Da xoa thong tin user");
-    }catch(ex){
+      await fs.promises.writeFile(globalConfig.deviceConfig, JSON.stringify(deviceConfig));
+    }
+    return deviceConfig;
+  },
+  async removeConfig(isRemoveUser) {
+    try {
+      if (isRemoveUser && await exists(globalConfig.configPath)) {
+        await fs.unlinkSync(globalConfig.configPath)
+        console.log("Da xoa thong tin user");
+      }
+
+      if (await exists(globalConfig.deviceConfig)) {
+        await fs.unlinkSync(globalConfig.deviceConfig)
+        console.log("Da xoa thong tin thiet bi");
+      }
+
+    } catch (ex) {
       console.error(ex);
     }
 
   },
-  getDeviceIP : () => {
+  getDeviceIP() {
     return getIp();
   },
-  getUserSSH: async() => {
+  async getUserSSH() {
     const { stdout, error } = await util.promisify(exec)("whoami");
     if (error) {
       return error;
     }
     return stdout.trim();
   },
-  delay : async (delayInms) => {
+  async getCpuUse() {
+    const cpuPercentage = await osu.cpu.usage();
+    return cpuPercentage;
+  },
+  getCpuCores() {
+    return osu.cpu.count()
+  },
+  async delay(delayInms) {
     return new Promise(resolve => setTimeout(resolve, delayInms));
   }
 }
